Angular Momenta

Orbital angular momentum

AngularMomentumAlgebra.OrbitalAngularMomentum — Type

OrbitalAngularMomentum()

The angular momentum $\tensor{L}$ is associated with the spatial coordinates $\theta$ and $\phi$.

AngularMomentumAlgebra.rme — Method

rme(ℓ′, 𝐋̂, ℓ)

Calculate the reduced matrix element of the orbital angular momentum:

\[\begin{equation} \tag{V13.2.67} \redmatrixel{\ell'}{\tensor{L}}{\ell} \defd \delta_{\ell'\ell}\sqrt{\ell(\ell+1)(2\ell+1)} \end{equation}\]

Examples

julia> rme(1, OrbitalAngularMomentum(), 1)
2.449489742783178

julia> rme(1, OrbitalAngularMomentum(), 2)
0

AngularMomentumAlgebra.couplings — Method

couplings(tensor::OrbitalAngularMomentum, ℓ)

Generate all quantum numbers ℓ′ for which ⟨ℓ′||::OrbitalAngularMomentum||ℓ⟩ does not vanish.

Spin angular momentum

AngularMomentumAlgebra.SpinAngularMomentum — Type

SpinAngularMomentum()

The spin angular momentum $\tensor{S}$ is the intrinsic angular momentum associated with the coordinate $s$.

AngularMomentumAlgebra.rme — Method

rme(s′, ::SpinAngularMomentum, s)

Calculate the reduced matrix element of the spin angular momentum:

\[\begin{equation} \tag{V13.2.95} \redmatrixel{s'}{\tensor{S}}{s} \defd \delta_{ss'} \sqrt{s(s+1)(2s+1)} \end{equation}\]

Examples

julia> rme(half(1), SpinAngularMomentum(), half(1))
1.224744871391589

julia> rme(half(1), SpinAngularMomentum(), half(3))
0

AngularMomentumAlgebra.couplings — Method

couplings(tensor::SpinAngularMomentum, s)

Generate all quantum numbers s′ for which ⟨s′||::SpinAngularMomentum||s⟩ does not vanish.

Total angular momentum

AngularMomentumAlgebra.TotalAngularMomentum — Type

TotalAngularMomentum()

The total angular momentum $\tensor{J} = \tensor{L} + \tensor{S}$ results from the coupling of the orbital and spin angular momenta.

AngularMomentumAlgebra.rme — Method

rme((ℓ′,s′,J′), ::TotalAngularMomentum, (ℓ,s,J))

Calculate the reduced matrix element of the total angular momentum:

\[\begin{equation} \tag{V13.2.38} \redmatrixel{\ell's'J'}{\tensor{J}}{\ell s J} \defd \delta_{\ell\ell'}\delta_{ss'}\delta_{JJ'} \sqrt{J(J+1)(2J+1)} \end{equation}\]

Examples

julia> rme((1,half(1),half(3)), TotalAngularMomentum(), (1,half(1),half(3)))
3.872983346207417

julia> rme((1,half(1),half(3)), TotalAngularMomentum(), (1,half(1),half(1)))
0

AngularMomentumAlgebra.couplings — Method

couplings(tensor::TotalAngularMomentum, (ℓ, s, J))

Generate all quantum numbers ℓ′s′J′ for which ⟨ℓ′s′J′||::TotalAngularMomentum||ℓsJ⟩ does not vanish.