Multipole expansions

Multipole expansions

AngularMomentumAlgebra.multipole_expand_scalar_product — Function.

multipole_expand_scalar_product(a, b, P, Q, c, d)

Multipole-expand the matrix element ⟨ab|P⋅Q|cd⟩, where the tensor P acts on orbitals a & c, and the tensor Q acts on orbitals b & d. The definition is taken from Eq. (13.1.26) of Varshalovich (1988).

source

AngularMomentumAlgebra.multipole_expand — Function.

multipole_expand(integral::OrbitalMatrixElement{2,A,<:CoulombInteraction,B})

Multipole-expand the two-body integral resulting from the Coulomb repulsion between two electrons.

source