Spherical Tensors

Spherical tensors

The spherical tensors are related to the spherical harmonics as

\[\tensor{C}^{(k)}_q \defd \sqrt{\frac{4\pi}{2k+1}} Y^k_q. \tag{V5.1.7}\]

AngularMomentumAlgebra.SphericalTensor — Type.

SphericalTensor(k)

Construct a spherical tensor of rank k.

source

The reduced matrix element of the spherical tensor is given by

\[\begin{aligned} \redmatrixel{\ell'}{\tensor{C}^{(k)}}{\ell} &= \angroot{\ell} C_{\ell 0;k,0}^{\ell'0} = (-)^{\ell-k} \angroot{\ell\ell'} \wignerthreej{\ell&k&\ell'\\0&0&0}. \end{aligned} \tag{V13.2.107}\]

For the reduced matrix element of the same spherical tensor in the $\ell s j$ basis, we must use the uncoupling formula together with the above formula in the uncoupled basis:

\[\begin{aligned} \redmatrixel{\ell' s' j'}{\tensor{C}^{(k)}}{\ell s j} &= \delta_{ss'}(-)^{j+\ell'+s+k} \angroot{jj'} \wignersixj{\ell&s&j\\j'&k&\ell'} \redmatrixel{\ell'}{\tensor{C}^{(k)}}{\ell}\\ &= (-)^{\ell'+\ell+s+j} \angroot{\ell\ell'jj'} \wignersixj{\ell&s&j\\j'&k&\ell'} \wignerthreej{\ell&k&\ell'\\0&0&0}. \end{aligned} \tag{V13.\{1.40,2.5\}}\]

AngularMomentumAlgebra.rme — Function.

rme(ℓ′,Cᵏ,ℓ)

Calculate the reduced matrix element ⟨ℓ′||C⁽ᵏ⁾||ℓ⟩ of the spherical tensor of rank k. Condon–Shortley phase convention and using the definition of Eq. (13.2.107) in Varshalovich (1988).

source

rme(o′, Cᵏ, o)

Calculate the reduced matrix element ⟨o′||C⁽ᵏ⁾||o⟩ of the spherical tensor of rank k, between the relativistic spin-orbitals o′ and o. Since the spherical tensors only act on the quantum numbers ℓ′ and ℓ, the reduced matrix has to be evaluated via the uncoupling formula given in Eqs. (13.1.40) & (13.2.5) in Varshalovich (1988), together with reduced matrix element for the uncoupled angular momenta ℓ′ and ℓ given by Eq. (13.2.107) (ibid).

source

AngularMomentumAlgebra.ranks — Function.

ranks(a, ::Type{SphericalTensor}, b)

Return which tensor ranks for spherical tensors that fulfill the triangle condition between spin-orbitals a and b.

source

Dipole operator

The dipole operator is a rank-1 Cartesian tensor that may be expressed using the rank-1 spherical tensor:

\[\hat{\vec{r}} \equiv \begin{bmatrix}\hat{x}\\\hat{y}\\\hat{z}\end{bmatrix} \equiv \begin{bmatrix} \frac{1}{\sqrt{2}}[-\tensor{C}^{(1)}_1 + \tensor{C}^{(1)}_{-1}]\\ \frac{\im}{\sqrt{2}}[\tensor{C}^{(1)}_1 + \tensor{C}^{(1)}_{-1}]\\ \tensor{C}^{(1)}_0 \end{bmatrix}\]

AngularMomentumAlgebra.Dipoles.𝐫̂ — Constant.

𝐫̂

The angular part of the dipole operator; the elements correspond to [x,y,z], i.e. the Cartesian tensor components. Can be entered as \bfr\hat.

Examples

julia> using AngularMomentumAlgebra.Dipoles

julia> z = 𝐫̂[3]
𝐂⁽¹⁾₀

julia> wigner_eckart(0, 0, z, 1, 0)
0.5773502691896256

source