Common routines

Common routines

AngularMomentumAlgebra.∏ — Function.

∏(ℓs...)

Calculates √((2ℓ₁+1)(2ℓ₂+1)...(2ℓₙ+1)), which is a common factor in angular momentum algebra.

source

AngularMomentumAlgebra.triangle_range — Function.

triangle_range(a,b)

Find all (even) k such that |a-b| ≤ k ≤ a + b. This is useful when expanding matrix elements of tensors between angular momenta a and b in multipoles k; triangle_range can then be used to decided which multipole terms are required.

source

AngularMomentumAlgebra.powneg1 — Function.

powneg1(k) = (-)ᵏ

Calculates powers of negative unity for integer k.

source

AngularMomentumAlgebra.jmⱼ — Function.

jmⱼ(o::SpinOrbital)

Return the angular momentum and its projection on the z axis of the spin-orbital o.

source

AngularMomentumAlgebra.spin — Function.

spin(o::SpinOrbital)

Return the spin of the spin-orbital o.

source

AngularMomentumAlgebra.LinearCombination — Type.

LinearCombination(Ts::Vector{T}, coeffs::Vector{N})

Represents a general linear combination of objects of type T.

source

AngularMomentumAlgebra.@linearly_combinable — Macro.

@linearly_combinable TT

Turns the type TT into a linearly combinable type, i.e. defines arithmetic operators.

Examples

julia> @linearly_combinable Symbol

julia> 4*(:x) - 5*(:y)
4 :x - 5 :y

source

AngularMomentumAlgebra.clebsch_gordan_condon_shortley — Function.

clebsch_gordan_condon_shortley(j₁, m₁, j₂, m₂, j₃, m₃=m₁+m₂)

Calculate the vector coupling coefficient ⟨j₁ m₁, j₂ m₂|j₃ m₃⟩ according to the Condon–Shortley phase convention and using the definition of Eq. (8.1.12) in Varshalovich (1988).

source