Common routines

Common routines

A commonly occurring factor in angular momentum algebra is

\[\begin{equation} \angroot{j_1j_2...j_n} \defd[(2j_1+1)(2j_2+1)...(2j_n+1)]^{1/2}. \tag{V13.1.3½} \end{equation}\]

AngularMomentumAlgebra.∏ — Function.

∏(ℓs...)

Calculates √((2ℓ₁+1)(2ℓ₂+1)...(2ℓₙ+1)), which is a common factor in angular momentum algebra.

source

AngularMomentumAlgebra.triangle_range — Function.

triangle_range(a,b)

Find all (even) k such that |a-b| ≤ k ≤ a + b. This is useful when expanding matrix elements of tensors between angular momenta a and b in multipoles k; triangle_range can then be used to decided which multipole terms are required.

source

AngularMomentumAlgebra.powneg1 — Function.

powneg1(k) = (-)ᵏ

Calculates powers of negative unity for integer k.

source

AngularMomentumAlgebra.jmⱼ — Function.

jmⱼ(o::SpinOrbital)

Return the angular momentum and its projection on the z axis of the spin-orbital o.

source

AngularMomentumAlgebra.spin — Function.

spin(o::SpinOrbital)

Return the spin of the spin-orbital o.

source