Coulomb interaction

Coulomb interaction

The Coulomb interaction between two electrons at coordinates $1$ and $2$, respectively, can in spherical coordinates be multipole-expanded as:

\[\begin{equation} \frac{1}{r_{12}} = \sum_{k=0}^\infty \frac{r_<^k}{r_>^{k+1}} P_k(\cos\theta) \end{equation}\]

which by the addition theorem for the spherical harmonics can be further expanded as

\[\begin{equation} \frac{1}{r_{12}} = \sum_{kq} \frac{4\pi}{2k+1} \frac{r_<^k}{r_>^{k+1}} \conj{{Y^k_q}}(1)Y^k_q(2)\equiv \sum_{kq} \frac{r_<^k}{r_>^{k+1}} \tensor{C}^k_q(1)\cdot\tensor{C}^k_q(2), \tag{V5.17.9} \end{equation}\]

where we in the last step have used the definition of the Spherical tensors.

AngularMomentumAlgebra.CoulombInteractionMultipole — Type.

CoulombInteractionMultipole(k, g)

Represents the kth multipole of the multipole expansion of the Coulomb interaction g.

source

AngularMomentumAlgebra.CoulombPotentialMultipole — Type.

CoulombPotentialMultipole

Type alias for contraction of the CoulombInteractionMultipole over one coordinate, thereby forming a potential in the other coordinate.

source