Tensors

Tensors

AngularMomentumAlgebra.Tensor — Type.

Tensor

Abstract base for any tensor.

source

AngularMomentumAlgebra.TensorComponent — Type.

TensorComponent(tensor, q)

Represents the qth component of a tensor; abs(q) ≤ rank(tensor).

source

AngularMomentumAlgebra.LinearCombinationTensor — Type.

LinearCombinationTensor

Represents a linear combination of tensor components.

source

The Wigner–Eckart theorem

The Wigner–Eckart theorem states that the matrix element of a tensor component $\tensor{T}^{(k)}_q$ can be evaluated as

\[\begin{equation} \matrixel{n'j'm'}{\tensor{T}^{(k)}_q}{njm}= (-)^{j'-m'} \begin{pmatrix} j'&k&j\\ -m'&\kappa&m \end{pmatrix} \redmatrixel{n'j'}{\tensor{T}^{(k)}_q}{nj}, \tag{V13.1.2} \end{equation}\]

where the reduced matrix element $\redmatrixel{n'j'}{\tensor{T}^{(k)}_q}{nj}$ does not depend on $m,m'$.

AngularMomentumAlgebra.wigner_eckart — Function.

wigner_eckart(j′, m′, T⁽ᵏ⁾q, j, m)

Computes the (spin-angular part of the) matrix element ⟨n′j′m′|Tᵏq|njm⟩, where T⁽ᵏ⁾q is the qth component of a tensor of rank k, using the definition of Eq. (13.1.1) in Varshalovich (1988).

source

Spherical tensors

The spherical tensors are related to the spherical harmonics as

\[\tensor{C}^{(k)}_q \defd \sqrt{\frac{4\pi}{2k+1}} Y^k_q. \tag{V5.1.7}\]

AngularMomentumAlgebra.SphericalTensor — Type.

SphericalTensor(k)

Construct a spherical tensor of rank k.

source

The reduced matrix element of the spherical tensor is given by

\[\begin{aligned} \redmatrixel{\ell'}{\tensor{C}^{(k)}}{\ell} &= \angroot{\ell} C_{\ell 0;k,0}^{\ell'0} = (-)^{\ell-k} \angroot{\ell\ell'} \begin{pmatrix} \ell&k&\ell'\\0&0&0 \end{pmatrix}. \end{aligned} \tag{V13.2.107}\]

AngularMomentumAlgebra.rme — Function.

rme(ℓ′,Cᵏ,ℓ)

Calculate the reduced matrix element ⟨ℓ′||C⁽ᵏ⁾||ℓ⟩ of the spherical tensor of rank k. Condon–Shortley phase convention and using the definition of Eq. (13.2.107) in Varshalovich (1988).

source