Tensors

Tensors

AngularMomentumAlgebra.Tensor — Type.

Tensor{k,label}

Abstract base for any tensor of rank k.

source

AngularMomentumAlgebra.TensorComponent — Type.

TensorComponent(tensor, q)

Represents the qth component of a tensor; abs(q) ≤ rank(tensor).

source

AngularMomentumAlgebra.LinearCombinationTensor — Type.

LinearCombinationTensor

Represents a linear combination of tensor components.

source

Product tensors

AngularMomentumAlgebra.TensorProduct — Type.

TensorProduct{K}(T, U)

A tensor of rank K formed from the product of the tensors T and U, according to

\[\begin{equation} \tag{V3.1.20} \tensor{X}^{(K)}_Q \equiv \{\tensor{T}^{(k_1)}\tensor{U}^{(k_2)}\}^{(K)}_Q \defd \tensor{T}^{(k_1)}_{q_1} \tensor{U}^{(k_2)}_{q_2} C_{k_1q_1k_2q_2}^{KQ} \end{equation}\]

source

Scalar product

The matrix element of scalar product of two tensors acting on different coordinates is given by

\[\begin{equation} \begin{aligned} &\matrixel{n_aj_am_a;n_bj_bm_b}{[\tensor{P}^{(k)}(1)\cdot\tensor{Q}^{(k)}(2)]}{n_cj_cm_c;n_dj_dm_d}\\ =& \frac{1}{\angroot{j_aj_b}} \sum_\alpha(-)^{-\alpha} C_{j_cm_c;k,\alpha}^{j_am_a} C_{j_dm_d;k,-\alpha}^{j_bm_b}\\ &\times \redmatrixel{n_aj_a}{\tensor{P}^{(k)}(1)}{n_cj_c} \redmatrixel{n_bj_b}{\tensor{Q}^{(k)}(2)}{n_dj_d} \end{aligned} \tag{V13.1.26} \end{equation}\]

Since the Clebsch–Gordan coefficients can be rewritten using 3j symbols and the 3j symbols vanish unless $m_c + \alpha - m_3 = m_d - \alpha - m_b = 0$, we have

\[\begin{equation} \alpha = m_a - m_c = m_d-m_b \implies -\alpha + m_a + m_b = m_b + m_c. \end{equation}\]

AngularMomentumAlgebra.TensorScalarProduct — Type.

TensorScalarProduct(T, U)

A tensor of rank 0 formed from the product of the tensors T and U (which have to have the same rank), according to

\[\begin{equation} \tag{V3.1.30,35} (\tensor{T}^{(k)} \cdot \tensor{U}^{(k)}) \defd (-)^k\angroot{k} \{\tensor{T}^{(k)} \tensor{U}^{(k)}\}^{(0)}_0 \equiv (-)^q \tensor{T}^{(k)}_{q} \tensor{U}^{(k)}_{-q} \end{equation}\]

source

LinearAlgebra.dot — Method.

dot(T::Tensor, U::Tensor)

Form the scalar product of the two tensors T and U, which need to have the same rank.

Examples

julia> SphericalTensor(4)⋅SphericalTensor(4)
(𝐂⁽⁴⁾⋅𝐂⁽⁴⁾)

source

AngularMomentumAlgebra.integrate_spinors — Method.

integrate_spinors((a,b), X, (c,d))

Perform the spin-angular integration of the scalar-product tensor X≡(T⁽ᵏ⁾⋅U⁽ᵏ⁾), where T acts on the coordinate of orbitals a & c and similarly, U acts on the coordinate of orbitals b & d, according to Eq. (13.1.26) of Varshalovich (1988).

source

The Wigner–Eckart theorem

The Wigner–Eckart theorem states that the matrix element of a tensor component $\tensor{T}^{(k)}_q$ can be evaluated as

\[\begin{equation} \matrixel{n'j'm'}{\tensor{T}^{(k)}_q}{njm}= (-)^{j'-m'} \begin{pmatrix} j'&k&j\\ -m'&q&m \end{pmatrix} \redmatrixel{n'j'}{\tensor{T}^{(k)}}{nj}, \tag{V13.1.2} \end{equation}\]

where the reduced matrix element $\redmatrixel{n'j'}{\tensor{T}^{(k)}}{nj}$ does not depend on $m,m'$.

AngularMomentumAlgebra.wigner_eckart — Function.

wigner_eckart(j′, m′, T⁽ᵏ⁾q, j, m)

Computes the (spin-angular part of the) matrix element ⟨n′j′m′|Tᵏq|njm⟩, where T⁽ᵏ⁾q is the qth component of a tensor of rank k, using the definition of Eq. (13.1.2) in Varshalovich (1988).

source

wigner_eckart(o′, T⁽ᵏ⁾q, o)

Computes the (spin-angular part of the) matrix element ⟨o′|Tᵏq|o⟩, where T⁽ᵏ⁾q is the qth component of a tensor of rank k, using the definition of Eq. (13.1.2) in Varshalovich (1988).

source