Tensors

AngularMomentumAlgebra.Tensor — Type

Tensor{k,label}

Abstract base for any tensor of rank k.

AngularMomentumAlgebra.TensorComponent — Type

TensorComponent(tensor, q)

Represents the qth component of a tensor; abs(q) ≤ rank(tensor).

AngularMomentumAlgebra.LinearCombinationTensor — Type

LinearCombinationTensor

Represents a linear combination of tensor components.

AngularMomentumAlgebra.system — Method

system(::Tensor)

A general tensor acts on the full system, i.e. all coordinates.

AngularMomentumAlgebra.complementary_space_factor — Function

complementary_space_factor(::SpatialSubSystems, k,
                           o′::SpinOrbital{<:Orbital},
                           o::SpinOrbital{<:Orbital})

Tensors acting on the spatial coordinates only, are diagonal in spin space:

\[\begin{equation} \tag{V13.2.3} \matrixel{n'\ell'm_\ell';s'm_s'}{\tensor{M}^{(k)}_q(r,\theta,\phi)}{n\ell m_\ell;sm_s} = \delta_{ss'}\delta{m_sm_s'} \matrixel{n'\ell'm_\ell'}{\tensor{M}^{(k)}_q(r,\theta,\phi)}{n\ell m_\ell} \end{equation}\]

complementary_space_factor(::SpinSubSystem, k,
                           o′::SpinOrbital{<:Orbital},
                           o::SpinOrbital{<:Orbital})

Tensors acting on the spin coordinates only, are diagonal in the spatial coordinates:

\[\begin{equation} \tag{V13.2.4} \matrixel{n'\ell'm_\ell';s'm_s'}{\tensor{N}^{(k)}_q(\xi)}{n\ell m_\ell;sm_s} = \delta_{\ell\ell'}\delta{m_\ell m_\ell'} \matrixel{s'm_s'}{\tensor{N}^{(k)}_q(\xi)}{sm_s} \end{equation}\]

complementary_space_factor(::SpatialSubSystems, k,
                           o′::SpinOrbital{<:RelativisticOrbital},
                           o::SpinOrbital{<:RelativisticOrbital})

Tensors acting on the spatial coordinates only, are diagonal in spin space; in the coupled basis, the following uncoupling formula must be employed:

\[\begin{equation} \tag{V13.2.5} \redmatrixel{n'\ell's'J'}{\tensor{M}^{(k)}}{n\ell s J} = \delta_{ss'} (-)^{J+\ell'+s'+k} \angroot{JJ'} \wignersixj{\ell&s&J\\J'&k&\ell'} \redmatrixel{n'\ell'}{\tensor{M}^{(k)}}{n\ell} \end{equation}\]

complementary_space_factor(::SpinSubSystem, k,
                           o′::SpinOrbital{<:RelativisticOrbital},
                           o::SpinOrbital{<:RelativisticOrbital})

Tensors acting on the spin coordinates only, are diagonal in the spatial coordinates; in the coupled basis, the following uncoupling formula must be employed:

\[\begin{equation} \tag{V13.2.6} \redmatrixel{n'\ell's'J'}{\tensor{N}^{(k)}}{n\ell s J} = \delta_{\ell\ell'} (-)^{J'+\ell+s+k} \angroot{JJ'} \wignersixj{s&\ell&J\\J'&k&s'} \redmatrixel{n's'}{\tensor{N}^{(k)}}{ns} \end{equation}\]

AngularMomentumAlgebra.rme_j′j — Function

rme_j′j(o′::SpinOrbital, T::Tensor, o::SpinOrbital)

Return the reduced matrix element ⟨o′||T||o⟩ and the two angular momenta pertaining to the tensor T, along with their projections.

Cartesian tensor components

The transformation of the tensor components of a rank-1 tensor from the Cartesian basis to the "natural" basis is given by:

\[\begin{equation} M(+1,0,-1 \leftarrow x, y, z)= \begin{bmatrix} -\frac{1}{\sqrt{2}}&\frac{\im}{\sqrt{2}}&\cdot\\ \cdot&\cdot&1\\ \frac{1}{\sqrt{2}}&\frac{\im}{\sqrt{2}}&\cdot \end{bmatrix} \end{equation}\]

AngularMomentumAlgebra.cartesian_tensor_component — Function

cartesian_tensor_component(t::Tensor{1}, c)

This returns the Cartesian tensor component c (valid choices are :x, :y, or :z) of the rank-1 tensor t, as a linear combination of its "natural" TensorComponents 1, 0, and -1. The transform matrix is M(+1, 0, -1 ← x, y, z) given in Table 1.2 on p. 14 of Varshalovich (1988).

Examples

julia> cartesian_tensor_component(Gradient(), :x)
- 0.707107 𝛁̂⁽¹⁾₁ + 0.707107 𝛁̂⁽¹⁾₋₁

Product tensors

AngularMomentumAlgebra.TensorProduct — Type

TensorProduct{K}(T, U)

A tensor of rank K formed from the product of the tensors T and U, according to

\[\begin{equation} \tag{V3.1.20} \tensor{X}^{(K)}_Q \equiv \{\tensor{T}^{(k_1)}\tensor{U}^{(k_2)}\}^{(K)}_Q \defd \tensor{T}^{(k_1)}_{q_1} \tensor{U}^{(k_2)}_{q_2} C_{k_1q_1k_2q_2}^{KQ} \end{equation}\]

Scalar product

The matrix element of scalar product of two tensors acting on different coordinates is given by

\[\begin{equation} \begin{aligned} &\matrixel{n_aj_am_a;n_bj_bm_b}{[\tensor{P}^{(k)}(1)\cdot\tensor{Q}^{(k)}(2)]}{n_cj_cm_c;n_dj_dm_d}\\ =& \frac{1}{\angroot{j_aj_b}} \sum_\alpha(-)^{-\alpha} C_{j_cm_c;k,\alpha}^{j_am_a} C_{j_dm_d;k,-\alpha}^{j_bm_b}\\ &\times \redmatrixel{n_aj_a}{\tensor{P}^{(k)}(1)}{n_cj_c} \redmatrixel{n_bj_b}{\tensor{Q}^{(k)}(2)}{n_dj_d} \end{aligned} \tag{V13.1.26} \end{equation}\]

Since the Clebsch–Gordan coefficients can be rewritten using 3j symbols and the 3j symbols vanish unless $m_c + \alpha - m_3 = m_d - \alpha - m_b = 0$, we have

\[\begin{equation} \alpha = m_a - m_c = m_d-m_b \implies -\alpha + m_a + m_b = m_b + m_c. \end{equation}\]

AngularMomentumAlgebra.TensorScalarProduct — Type

TensorScalarProduct(T, U)

A tensor of rank 0 formed from the product of the tensors T and U (which have to have the same rank), according to

\[\begin{equation} \tag{V3.1.30,35} (\tensor{T}^{(k)} \cdot \tensor{U}^{(k)}) \defd (-)^k\angroot{k} \{\tensor{T}^{(k)} \tensor{U}^{(k)}\}^{(0)}_0 \equiv (-)^q \tensor{T}^{(k)}_{q} \tensor{U}^{(k)}_{-q} \end{equation}\]

LinearAlgebra.dot — Method

dot(T::Tensor, U::Tensor)

Form the scalar product of the two tensors T and U, which need to have the same rank.

Examples

julia> SphericalTensor(4)⋅SphericalTensor(4)
(𝐂̂⁽⁴⁾⋅𝐂̂⁽⁴⁾)

AngularMomentumAlgebra.integrate_spinors — Method

integrate_spinors((a,b), X, (c,d))

Perform the spin-angular integration of the scalar-product tensor X≡(T⁽ᵏ⁾⋅U⁽ᵏ⁾), where T acts on the coordinate of orbitals a & c and similarly, U acts on the coordinate of orbitals b & d, according to Eq. (13.1.26) of Varshalovich (1988).

The Wigner–Eckart theorem

The Wigner–Eckart theorem states that the matrix element of a tensor component $\tensor{T}^{(k)}_q$ can be evaluated as

\[\begin{equation} \matrixel{n'j'm'}{\tensor{T}^{(k)}_q}{njm}= (-)^{j'-m'} \begin{pmatrix} j'&k&j\\ -m'&q&m \end{pmatrix} \redmatrixel{n'j'}{\tensor{T}^{(k)}}{nj}, \tag{V13.1.2} \end{equation}\]

where the reduced matrix element $\redmatrixel{n'j'}{\tensor{T}^{(k)}}{nj}$ does not depend on $m,m'$.

AngularMomentumAlgebra.wigner_eckart — Function

wigner_eckart(o′::SpinOrbital, Tᵏq::TensorComponent, o::SpinOrbital)

Calculate the matrix element ⟨o′|Tᵏq|o⟩ via Wigner–Eckart's theorem, automatically applying the appropriate uncoupling formulas (V13.2.5,6) of Varshalovich (1988) if o′ and o are coupled spin-orbitals and Tᵏ acts on one part of the system only.

LinearAlgebra.dot — Method

dot(o′, T, o)

Calculates the matrix element ⟨o′|T|o⟩ using wigner_eckart.