Tensors

Tensors

AngularMomentumAlgebra.Tensor — Type.

Tensor

Abstract base for any tensor.

source

Spherical tensors

The spherical tensors are related to the spherical harmonics as

\[\begin{equation} \tensor{C}^{(k)}_q \defd \sqrt{\frac{4\pi}{2k+1}} Y^k_q. \tag{V5.1.7} \end{equation}\]

AngularMomentumAlgebra.SphericalTensor — Type.

SphericalTensor(k)

Construct a spherical tensor of rank k.

source

The reduced matrix element of the spherical tensor is given by

\[\begin{equation} \begin{aligned} \redmatrixel{\ell'}{\tensor{C}^{(k)}}{\ell} &= \angroot{\ell} C_{\ell 0;k,0}^{\ell'0} = (-)^{\ell-k} \angroot{\ell\ell'} \begin{pmatrix} \ell&k&\ell'\\0&0&0 \end{pmatrix}. \end{aligned} \tag{V13.2.107} \end{equation}\]

AngularMomentumAlgebra.rme — Function.

rme(ℓ′,Cᵏ,ℓ)

Calculate the reduced matrix element ⟨ℓ′||C⁽ᵏ⁾||ℓ⟩ of the spherical tensor of rank k. Condon–Shortley phase convention and using the definition of Eq. (13.2.107) in Varshalovich (1988).

source